refactor: split constant.rs into a multiple files in a folder and move structs under models
