+++ src/web/constant/app.rs
@@ -1,0 +1,8 @@
+// Social Media URLs
+pub const DISCORD_URL: &str = "https://discord.gg/HjaGsBBtFg";
+pub const GITHUB_URL: &str = "https://github.com/autumn-order";
+
+// Websites
+
+pub const BLACK_ROSE_WEBSITE_URL: &str = "https://black-rose.space";
+pub const APPLICATIONS_URL: &str = "https://apply.autumn-order.com";
